KotH Radius Added
Added King of the Hill radius that awards bonus points to whomever stands in it (currently 1.5x the nornal pointsPerInterval) for As a player, I want a King of the Hill capture section on the map to add diversity to gameplay.
Also added slight luminance to player-coloured tiles for As a player, I want tiles to light up in my colour when I capture them.