Player Lobby UI Revamp
Altered the player lobby, so that players may now select from different models and starting weapons before entering the game, as well as this the lobby uses a ready-up system instead of always starting on the press of any "Start" key. For As a player, I want to select my team color from the lobby menu so that I can quickly identify my team in-game,
As a player, I want to select my weapon type from the lobby so that I can tailor gameplay to what I enjoy, and
As a player, I want to be able to select from a variety of models for my player character from the lobby menu so that it’s easier to identify who I am while playing. (2 hours)